frontend: Split Settings page into tabbed sub-pages
The settings page had grown into a single ~560-line component mixing
account info, profile, timezone, sustaining membership, MFA devices,
and API keys. Split it into Profile, Security, API Keys, and Billing
tabs, each in its own component, so future settings features have a
natural place to live instead of growing the monolith further.

Each tab only mounts (and fetches its data) when selected, instead of
all four fetching eagerly on page load as before.

Closes #256.
